PR: Refactor Campaign Options
